Добавлена поддержка генерации массивов больше, чем 1000 элементов. Небольшие исправления ошибок.

+++ ProtocolLibrary/include/PackageTypeForClient.h
@@ -5,10 +5,10 @@
 
 struct PackageTypeToClient {
     QString valueType;
-    quint64 bytes;
+    quint32 bytes;
 
     PackageTypeToClient() : valueType("int"), bytes(0) {}
-    PackageTypeToClient(QString _valueType, quint64 _bytes) : valueType(_valueType), bytes(_bytes) {}
+    PackageTypeToClient(QString _valueType, quint32 _bytes) : valueType(_valueType), bytes(_bytes) {}
 };
 
 inline QDataStream &operator>>(QDataStream &in, PackageTypeToClient& packageTypeToClient) {
